backend/requirements.txt
# Quantum Navigator v2.0 - Backend Requirements
# ==============================================

# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Quantum Computing - Neutral Atoms
pulser-core>=0.18.0
pulser-simulation>=0.18.0

# Quantum Computing - General
qiskit>=1.2.0
qiskit-aer>=0.14.0

# QEC Simulation
stim>=1.13.0
pymatching>=2.0.0

# Scientific Computing
numpy>=1.26.0
scipy>=1.12.0

# Optional: Tensor Network backends (for large neutral atom simulations)
# quimb>=1.8.0  # Uncomment for TN support

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
